Dashboard refonte: fix user permissions qui n'affiche plus les dossier dans la side bar quand on refresh
